Refactor color components
