+++ apps/studio/src/routes/editor/LayersPanel/BrandTab/ColorPanel/index.tsx
@@ -1,33 +1,14 @@
+import { useProjectsManager } from '@/components/Context';
+import { invokeMainChannel } from '@/lib/utils';
+import { MainChannels } from '@onlook/models/constants';
 import { Button } from '@onlook/ui/button';
 import { Icons } from '@onlook/ui/icons';
+import { cn } from '@onlook/ui/utils';
+import { Color } from '@onlook/utility';
 import { observer } from 'mobx-react-lite';
 import { useEffect, useState } from 'react';
-import { useProjectsManager } from '@/components/Context';
-import { MainChannels } from '@onlook/models/constants';
-import { invokeMainChannel } from '@/lib/utils';
-import { Color } from '@onlook/utility';
-import { cn } from '@onlook/ui/utils';
 import { BrandPalletGroup, type ColorItem } from './ColorPalletGroup';
-
-interface ColorRowProps {
-    label: string;
-    colors: string[];
-}
-
-const ColorRow = ({ label, colors }: ColorRowProps) => (
-    <div className="flex flex-col gap-1">
-        <span className="text-xs text-muted-foreground">{label}</span>
-        <div className="grid grid-cols-6 gap-1">
-            {colors.map((color, index) => (
-                <div
-                    key={`${label}-${index}`}
-                    className="w-full aspect-square rounded-lg cursor-pointer hover:ring-2 hover:ring-border-primary border border-white/10"
-                    style={{ backgroundColor: color }}
-                />
-            ))}
-        </div>
-    </div>
-);
+import { ColorRow } from './ColorRow';
 
 interface ThemeColors {
     [key: string]: string;
